requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
Pillow==9.5.0
reportlab==4.0.4
Werkzeug==2.3.6
PyMuPDF==1.22.5
python-docx==0.8.11
python-pptx==0.6.21
openpyxl==3.1.2
gunicorn==21.2.0